Implement fixes to track development
The projects has progressed quite a bit, thus this commit adds a number
of fixes to stay compatible and robust.

Changes (highlights)
- Adds some more timings for easier debugging of reset timings
- Fixes a bug where mkcert certificates where placed in the wrong directory.
- Separate traefik reset from the rest to make it easier to test the
  setup while another process has hold of port 80+443
- Stop cloning templates into display-client as the repo instructions
  no longer requires this
- Copy in example-access-config.json as access-config.json in display-admin
  to shadow over a bug
- Drop api and nginx-api from our overrides

+++ scripts/admin-reset.sh
@@ -8,8 +8,10 @@
 set -euo pipefail
 IFS=$'\n\t'
 
-docker compose down || true
-cp ../../etc/admin_config.json ./public/config.json
+docker compose down -v || true
+test -f ./public/config.json || cp ../../etc/admin_config.json ./public/config.json
+test -f ./public/access-config.json || cp ./public/example-access-config.json ./public/access-config.json
+
 cp ../../etc/admin_docker-compose.override.yml ./docker-compose.override.yml
 
 docker compose up -d
